added greeting

+++ events/message.js
@@ -25,10 +25,10 @@ async function handleCommands(client, message) {
             await message.channel.send(ccmd.response);
             return true;
         } else {
-            const alias = Alias.findOne({ where: { name: command } });
+            const alias = await Alias.findOne({ where: { name: command } });
             if (!alias) return false;
             if (alias.type === "ccmd") {
-                const accmd = await CustomCommand.findOne({ where: { commandName: command } });
+                const accmd = await CustomCommand.findOne({ where: { commandName: alias.command } });
                 if (accmd) {
                     await message.channel.send(accmd.response);
                     return true;
@@ -60,7 +60,7 @@ function handleModCommands(client, message) {
     if (cmd) {
         if (cmd.perm <= permissions.mod) return cmd.run(client, message, args);
         if (cmd.perm <= permissions.dev && message.member.roles.cache.has(client.config.roles.dev)) return cmd.run(client, message, args);
-        if (message.author.id == message.guild.ownerId) return cmd.run(client, message, args);
+        if (message.author.id === message.guild.ownerId) return cmd.run(client, message, args);
     }
 }
 
@@ -76,20 +76,21 @@ function unverify(client, member) {
  * @param  {import("discord.js").Message} message
  * @returns {boolean}
  */
-function checkMessage(client, message) {
+export function checkMessage(client, message) {
     if (!message.guild) return false;
     if (message.author.id === client.user.id) return false;
-    if (message.member && message.member.roles && message.member.roles.cache.hasAny(client.config.roles.noFilter)) return false;
-    const links = [...message.content.matchAll(linkRegex)];
+    if (message.member && message.member.roles && message.member.roles.cache.map((r) => r.id).some((r) => client.config.roles.noFilter.includes(r))) return false;
+    let links = [...message.content.matchAll(linkRegex)];
+    if (message.channel.id === client.config.channels.clips) links = links.filter((l) => !l.toString().replace(/^(http|https):\/\//, "").startsWith("clips.twitch.tv"));
     if (links.length > 0) {
-        message.delete();
+        if (message.deletable) message.delete();
         let previousLinks = client.linkUsage[message.author.id];
         if (!previousLinks) previousLinks = [];
         previousLinks = previousLinks.filter((link) => message.createdAt.getTime() - link.ts < 5000);
         const newLinks = previousLinks;
         let shouldUnverify = false;
         links.forEach((link) => {
-            if (previousLinks.some((previousLink) => previousLink.url == link)) shouldUnverify = true;
+            if (previousLinks.some((previousLink) => previousLink.url === link)) shouldUnverify = true;
             newLinks.push({ ts: message.createdAt.getTime(), url: link });
         });
         if (shouldUnverify && message.member.roles.cache.has(client.config.roles.verify)) unverify(client, message.member);
@@ -98,12 +99,12 @@ function checkMessage(client, message) {
     if (client.blacklist.length > 0) {
         const lowerMessage = message.content.toLowerCase();
         if (client.blacklist.some((blword) => lowerMessage.indexOf(blword) !== -1)) {
-            message.delete().catch((e) => logger.error("Received an error deleting a message:\n" + e.stack));
+            if (message.deletable) message.delete().catch((e) => logger.error("Received an error deleting a message:\n" + e.stack));
             message.author.createDM().then((dmChannel) => {
                 dmChannel.send(`Ihre Nachricht mit dem Inhalt **${message.content.slice(0, 1900)}** wurde entfernt. Melden Sie sich bei Fragen an einen Moderator.`);
             });
+            return true;
         }
-        return true;
     }
     return false;
 }
@@ -113,20 +114,27 @@ function checkMessage(client, message) {
  * @param  {import("discord.js").Message} message
  */
 export async function handle(client, message) {
+    logger.silly("message received");
     if (!message) return;
+    logger.silly("message exists");
     if (message.partial) message = await message.fetch();
+    logger.silly("fetched possible partial message");
     if (checkMessage(client, message)) return;
+    logger.silly("message was clean.");
     if (Math.random() > 0.999) message.react(client.config.emojis.randomReaction);
-    if (message.channelId == client.config.channels.verificate && message.content.toLowerCase == "!zz") {
+    logger.silly("checking for verification");
+    if (message.channelId === client.config.channels.verificate && message.content.toLowerCase() === `${client.config.prefix}zz`) {
+        logger.silly("verifying user..");
         message.member.roles.add(client.config.roles.verify);
         verify(client, message.author.username);
-        message.delete();
+        if (message.deletable) message.delete();
         return;
     }
     if (client.config.channels.commands.includes(message.channelId)) return handleCommands(client, message);
 
     if (client.config.channels.adminCommands.includes(message.channelId)) return handleModCommands(client, message);
     if (!message.guild) return;
+    if (message.author.bot) return;
     const [xpuser] = await XPUser.findOrCreate({
         where: {
             discordId: message.author.id,
@@ -139,7 +147,7 @@ export async function handle(client, message) {
         },
     });
     if (xpuser.username !== message.author.username) await xpuser.update({ username: message.author.username });
-    const exp = Math.max(10, Math.floor(((message.content.length) - 2) / 5));
+    const exp = Math.min(10, Math.floor(((message.content.length) - 2) / 5));
     await xpuser.increment("experience", {
         by: exp,
     });
